Repackaged Condition implementations
