+++ src/main/java/dev/qadenz/automation/conditions/impl/CssPropertyOfElement.java
@@ -7,9 +7,12 @@ A copy of the License may be obtained at
 
 https://polyformproject.org/licenses/internal-use/1.0.0/
  */
-package dev.qadenz.automation.conditions;
+package dev.qadenz.automation.conditions.impl;
 
 import dev.qadenz.automation.commands.WebInspector;
+import dev.qadenz.automation.conditions.Condition;
+import dev.qadenz.automation.conditions.Conditions;
+import dev.qadenz.automation.conditions.Expectation;
 import dev.qadenz.automation.ui.Locator;
 
 /**
