+++ src/main/java/dev/qadenz/automation/conditions/impl/DirectTextOfElement.java
@@ -7,37 +7,41 @@ A copy of the License may be obtained at
 
 https://polyformproject.org/licenses/internal-use/1.0.0/
  */
-package dev.qadenz.automation.conditions;
+package dev.qadenz.automation.conditions.impl;
 
 import dev.qadenz.automation.commands.WebInspector;
+import dev.qadenz.automation.conditions.Condition;
+import dev.qadenz.automation.conditions.Conditions;
+import dev.qadenz.automation.conditions.Expectation;
 import dev.qadenz.automation.ui.Locator;
 
 /**
- * A Condition to evaluate the visible inner text of an element.
+ * A Condition to evaluate the visible inner text of an element, excluding the text of any descendant elements on the
+ * DOM.
  *
  * @author Tim Slifer
  */
-public class TextOfElement implements Condition {
+public class DirectTextOfElement implements Condition {
     
     private Locator locator;
     private Expectation<String> expectation;
     
     private String elementText;
     
-    public TextOfElement(Locator locator, Expectation<String> expectation) {
+    public DirectTextOfElement(Locator locator, Expectation<String> expectation) {
         this.locator = locator;
         this.expectation = expectation;
     }
     
     @Override
     public String description() {
-        return "Text of element [" + locator.getName() + "] " + expectation.description() + ".";
+        return "Direct text of element [" + locator.getName() + "] " + expectation.description() + ".";
     }
     
     @Override
     public Boolean result() {
         WebInspector webInspector = new WebInspector(Conditions.class);
-        elementText = webInspector.getTextOfElement(locator);
+        elementText = webInspector.getDirectTextOfElement(locator);
         
         return expectation.matcher().matches(elementText);
     }
