+++ src/main/java/dev/qadenz/automation/conditions/impl/DirectTextOfElementAsDate.java
@@ -7,31 +7,34 @@ A copy of the License may be obtained at
 
 https://polyformproject.org/licenses/internal-use/1.0.0/
  */
-package dev.qadenz.automation.conditions;
+package dev.qadenz.automation.conditions.impl;
 
 import dev.qadenz.automation.commands.WebInspector;
+import dev.qadenz.automation.conditions.Condition;
+import dev.qadenz.automation.conditions.Conditions;
+import dev.qadenz.automation.conditions.TemporalExpectation;
 import dev.qadenz.automation.ui.Locator;
 
-import java.time.LocalTime;
+import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 
 /**
  * A Condition to evaluate the visible inner text of an element, excluding the text of any descendant elements on the
- * DOM, as a formatted LocalTime.
+ * DOM, as a formatted LocalDate.
  *
  * @author Tim Slifer
  */
-public class DirectTextOfElementAsTime implements Condition {
+public class DirectTextOfElementAsDate implements Condition {
     
     private Locator locator;
     private DateTimeFormatter dateTimeFormatter;
-    private TemporalExpectation<LocalTime> expectation;
+    private TemporalExpectation<LocalDate> expectation;
     
     private String elementText;
-    private LocalTime elementTime;
+    private LocalDate elementDate;
     
-    public DirectTextOfElementAsTime(Locator locator, DateTimeFormatter dateTimeFormatter,
-            TemporalExpectation<LocalTime> expectation) {
+    public DirectTextOfElementAsDate(Locator locator, DateTimeFormatter dateTimeFormatter,
+            TemporalExpectation<LocalDate> expectation) {
         this.locator = locator;
         this.dateTimeFormatter = dateTimeFormatter;
         this.expectation = expectation;
@@ -41,20 +44,20 @@ public class DirectTextOfElementAsTime implements Condition {
     
     @Override
     public String description() {
-        return "Direct text of element [" + locator.getName() + "] as LocalTime " + expectation.description() + ".";
+        return "Direct text of element [" + locator.getName() + "] as LocalDate " + expectation.description() + ".";
     }
     
     @Override
     public Boolean result() {
         WebInspector webInspector = new WebInspector(Conditions.class);
         elementText = webInspector.getDirectTextOfElement(locator);
-        elementTime = webInspector.getDirectTextOfElementAsTime(locator, dateTimeFormatter);
+        elementDate = webInspector.getDirectTextOfElementAsDate(locator, dateTimeFormatter);
         
-        return expectation.matcher().matches(elementTime);
+        return expectation.matcher().matches(elementDate);
     }
     
     @Override
     public String output() {
-        return "Found [" + elementText + "] formatted as [" + dateTimeFormatter.format(elementTime) + "].";
+        return "Found [" + elementText + "] formatted as [" + dateTimeFormatter.format(elementDate) + "].";
     }
 }
