+++ src/main/java/dev/qadenz/automation/conditions/impl/DirectTextOfElementAsDateTime.java
@@ -7,20 +7,24 @@ A copy of the License may be obtained at
 
 https://polyformproject.org/licenses/internal-use/1.0.0/
  */
-package dev.qadenz.automation.conditions;
+package dev.qadenz.automation.conditions.impl;
 
 import dev.qadenz.automation.commands.WebInspector;
+import dev.qadenz.automation.conditions.Condition;
+import dev.qadenz.automation.conditions.Conditions;
+import dev.qadenz.automation.conditions.TemporalExpectation;
 import dev.qadenz.automation.ui.Locator;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
 /**
- * A Condition to evaluate the visible inner text of an element as a formatted LocalDateTime.
+ * A Condition to evaluate the visible inner text of an element, excluding the text of any descendant elements on the
+ * DOM, as a formatted LocalDateTime.
  *
  * @author Tim Slifer
  */
-public class TextOfElementAsDateTime implements Condition {
+public class DirectTextOfElementAsDateTime implements Condition {
     
     private Locator locator;
     private DateTimeFormatter dateTimeFormatter;
@@ -29,7 +33,7 @@ public class TextOfElementAsDateTime implements Condition {
     private String elementText;
     private LocalDateTime elementDateTime;
     
-    public TextOfElementAsDateTime(Locator locator, DateTimeFormatter dateTimeFormatter,
+    public DirectTextOfElementAsDateTime(Locator locator, DateTimeFormatter dateTimeFormatter,
             TemporalExpectation<LocalDateTime> expectation) {
         this.locator = locator;
         this.dateTimeFormatter = dateTimeFormatter;
@@ -40,14 +44,14 @@ public class TextOfElementAsDateTime implements Condition {
     
     @Override
     public String description() {
-        return "Text of element [" + locator.getName() + "] as LocalDateTime " + expectation.description() + ".";
+        return "Direct text of element [" + locator.getName() + "] as LocalDateTime" + expectation.description() + ".";
     }
     
     @Override
     public Boolean result() {
         WebInspector webInspector = new WebInspector(Conditions.class);
-        elementText = webInspector.getTextOfElement(locator);
-        elementDateTime = webInspector.getTextOfElementAsDateTime(locator, dateTimeFormatter);
+        elementText = webInspector.getDirectTextOfElement(locator);
+        elementDateTime = webInspector.getDirectTextOfElementAsDateTime(locator, dateTimeFormatter);
         
         return expectation.matcher().matches(elementDateTime);
     }
