+++ src/main/java/dev/qadenz/automation/conditions/impl/DirectTextOfElementAsTime.java
@@ -7,31 +7,34 @@ A copy of the License may be obtained at
 
 https://polyformproject.org/licenses/internal-use/1.0.0/
  */
-package dev.qadenz.automation.conditions;
+package dev.qadenz.automation.conditions.impl;
 
 import dev.qadenz.automation.commands.WebInspector;
+import dev.qadenz.automation.conditions.Condition;
+import dev.qadenz.automation.conditions.Conditions;
+import dev.qadenz.automation.conditions.TemporalExpectation;
 import dev.qadenz.automation.ui.Locator;
 
-import java.time.LocalDateTime;
+import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
 
 /**
  * A Condition to evaluate the visible inner text of an element, excluding the text of any descendant elements on the
- * DOM, as a formatted LocalDateTime.
+ * DOM, as a formatted LocalTime.
  *
  * @author Tim Slifer
  */
-public class DirectTextOfElementAsDateTime implements Condition {
+public class DirectTextOfElementAsTime implements Condition {
     
     private Locator locator;
     private DateTimeFormatter dateTimeFormatter;
-    private TemporalExpectation<LocalDateTime> expectation;
+    private TemporalExpectation<LocalTime> expectation;
     
     private String elementText;
-    private LocalDateTime elementDateTime;
+    private LocalTime elementTime;
     
-    public DirectTextOfElementAsDateTime(Locator locator, DateTimeFormatter dateTimeFormatter,
-            TemporalExpectation<LocalDateTime> expectation) {
+    public DirectTextOfElementAsTime(Locator locator, DateTimeFormatter dateTimeFormatter,
+            TemporalExpectation<LocalTime> expectation) {
         this.locator = locator;
         this.dateTimeFormatter = dateTimeFormatter;
         this.expectation = expectation;
@@ -41,20 +44,20 @@ public class DirectTextOfElementAsDateTime implements Condition {
     
     @Override
     public String description() {
-        return "Direct text of element [" + locator.getName() + "] as LocalDateTime" + expectation.description() + ".";
+        return "Direct text of element [" + locator.getName() + "] as LocalTime " + expectation.description() + ".";
     }
     
     @Override
     public Boolean result() {
         WebInspector webInspector = new WebInspector(Conditions.class);
         elementText = webInspector.getDirectTextOfElement(locator);
-        elementDateTime = webInspector.getDirectTextOfElementAsDateTime(locator, dateTimeFormatter);
+        elementTime = webInspector.getDirectTextOfElementAsTime(locator, dateTimeFormatter);
         
-        return expectation.matcher().matches(elementDateTime);
+        return expectation.matcher().matches(elementTime);
     }
     
     @Override
     public String output() {
-        return "Found [" + elementText + "] formatted as [" + dateTimeFormatter.format(elementDateTime) + "].";
+        return "Found [" + elementText + "] formatted as [" + dateTimeFormatter.format(elementTime) + "].";
     }
 }
