+++ src/main/java/dev/qadenz/automation/conditions/impl/PresenceOfElement.java
@@ -7,8 +7,10 @@ A copy of the License may be obtained at
 
 https://polyformproject.org/licenses/internal-use/1.0.0/
  */
-package dev.qadenz.automation.conditions;
+package dev.qadenz.automation.conditions.impl;
 
+import dev.qadenz.automation.conditions.Condition;
+import dev.qadenz.automation.conditions.Expectation;
 import dev.qadenz.automation.ui.Locator;
 import dev.qadenz.automation.ui.WebFinder;
