+++ src/main/java/dev/qadenz/automation/conditions/impl/SelectedMenuOption.java
@@ -7,44 +7,46 @@ A copy of the License may be obtained at
 
 https://polyformproject.org/licenses/internal-use/1.0.0/
  */
-package dev.qadenz.automation.conditions;
+package dev.qadenz.automation.conditions.impl;
 
 import dev.qadenz.automation.commands.WebInspector;
+import dev.qadenz.automation.conditions.Condition;
+import dev.qadenz.automation.conditions.Conditions;
+import dev.qadenz.automation.conditions.Expectation;
 import dev.qadenz.automation.ui.Locator;
 
 /**
- * A Condition to evaluate the visible inner text of an element, excluding the text of any descendant elements on the
- * DOM.
+ * A Condition for evaluating the currently selected option of a Select menu element.
  *
  * @author Tim Slifer
  */
-public class DirectTextOfElement implements Condition {
+public class SelectedMenuOption implements Condition {
     
     private Locator locator;
     private Expectation<String> expectation;
     
-    private String elementText;
+    private String selectedOption;
     
-    public DirectTextOfElement(Locator locator, Expectation<String> expectation) {
+    public SelectedMenuOption(Locator locator, Expectation<String> expectation) {
         this.locator = locator;
         this.expectation = expectation;
     }
     
     @Override
     public String description() {
-        return "Direct text of element [" + locator.getName() + "] " + expectation.description() + ".";
+        return "Selected option of menu element [" + locator.getName() + "] " + expectation.description() + ".";
     }
     
     @Override
     public Boolean result() {
         WebInspector webInspector = new WebInspector(Conditions.class);
-        elementText = webInspector.getDirectTextOfElement(locator);
+        selectedOption = webInspector.getSelectedMenuOption(locator);
         
-        return expectation.matcher().matches(elementText);
+        return expectation.matcher().matches(selectedOption);
     }
     
     @Override
     public String output() {
-        return "Found [" + elementText + "].";
+        return "Found [" + selectedOption + "].";
     }
 }
