+++ src/main/java/dev/qadenz/automation/conditions/impl/TextOfElements.java
@@ -7,45 +7,51 @@ A copy of the License may be obtained at
 
 https://polyformproject.org/licenses/internal-use/1.0.0/
  */
-package dev.qadenz.automation.conditions;
+package dev.qadenz.automation.conditions.impl;
 
 import dev.qadenz.automation.commands.WebInspector;
+import dev.qadenz.automation.conditions.Condition;
+import dev.qadenz.automation.conditions.Conditions;
+import dev.qadenz.automation.conditions.Expectation;
 import dev.qadenz.automation.ui.Locator;
-import dev.qadenz.automation.ui.LocatorGroup;
+
+import java.util.List;
 
 /**
- * A Condition for evaluating whether a group of elements is enabled.
+ * A Condition to evaluate the visible inner text of each instance of an element.
  *
  * @author Tim Slifer
  */
-public class EnabledStateOfElements implements Condition {
+public class TextOfElements implements Condition {
     
-    private LocatorGroup locatorGroup;
-    private Expectation<Boolean> expectation;
+    private Locator locator;
+    private Expectation<String> expectation;
     
+    private List<String> elementValues;
     private StringBuilder failures = new StringBuilder();
     
-    public EnabledStateOfElements(LocatorGroup locatorGroup, Expectation<Boolean> expectation) {
-        this.locatorGroup = locatorGroup;
+    public TextOfElements(Locator locator, Expectation<String> expectation) {
+        this.locator = locator;
         this.expectation = expectation;
     }
     
     @Override
     public String description() {
-        return "Enabled state of elements [" + locatorGroup.getName() + "] " + expectation.description() + ".";
+        return "Text of each instance of element [" + locator.getName() + "] " + expectation.description() + ".";
     }
     
     @Override
     public Boolean result() {
         Boolean match = null;
         WebInspector webInspector = new WebInspector(Conditions.class);
+        elementValues = webInspector.getTextOfElements(locator);
         
-        for (Locator locator : locatorGroup) {
-            boolean enabled = webInspector.getEnabledStateOfElement(locator);
-            boolean instanceMatch = expectation.matcher().matches(enabled);
+        for (int i = 0; i < elementValues.size(); i++) {
+            String instanceValue = elementValues.get(i);
+            boolean instanceMatch = expectation.matcher().matches(instanceValue);
             
             if (!instanceMatch) {
-                failures.append("--> Element [" + locator.getName() + "] was [" + enabled + "].\n");
+                failures.append("--> at index [" + i + "], found [" + instanceValue + "].\n");
             }
             
             if (match == null || match) {
